+++ frontend/src/components/SearchBar/Calendar/MonthlyCalendar.tsx
@@ -3,6 +3,7 @@ import { useState } from "react";
 import { calendarModalState, calendarFilterState } from "recoil/Atoms";
 import * as S from "components/SearchBar/Calendar/CalendarStyles";
 import DateBlock from "./DateBlock";
+import { daysOfWeek, getDateList, getCalendarMonth } from "util/Calendar";
 interface MonthProps {
   monthType: string;
 }
@@ -13,44 +14,16 @@ const MonthlyCalendar: React.FunctionComponent<MonthProps> = ({
   const { year, month, nextMonth, today } = useRecoilValue(calendarModalState);
   const [checkInOut, setCheckInOut] = useRecoilState(calendarFilterState);
   const { checkIn, checkOut } = checkInOut;
-
-  let currentMonth: number = 0;
-
-  const getCalendarMonth = (type: string) => {
-    switch (type) {
-      case "PREVIOUS":
-        currentMonth = month - 1;
-        break;
-      case "CURRENT":
-        currentMonth = month;
-        break;
-      case "NEXT":
-        currentMonth = month + 1;
-        break;
-      case "MONTH_AFTER_NEXT":
-        currentMonth = month + 2;
-        break;
-      default:
-        console.error("Unhandled month type! Check the type of the month!");
-    }
-  };
+  const currentMonth = getCalendarMonth(monthType, month);
+  const dates = getDateList(year, currentMonth);
   const checkValidDate = (date: number) => {
     const calendarDate = new Date(year, currentMonth, date);
     return !date || calendarDate < new Date() ? "invalid-date" : "valid-date";
   };
 
-  getCalendarMonth(monthType);
+  const realMonth = new Date(year, currentMonth).getMonth() + 1;
+  const realYear = new Date(year, currentMonth).getFullYear();
 
-  let realMonth = new Date(year, currentMonth).getMonth() + 1;
-  let realYear = new Date(year, currentMonth).getFullYear();
-
-  const daysOfWeek: string[] = ["일", "월", "화", "수", "목", "금", "토"];
-  const startDay: number = new Date(year, currentMonth, 1).getDay();
-  const endDate: number = new Date(year, currentMonth + 1, 0).getDate();
-  const emptyDates: number[] = Array(startDay).fill("");
-  const filledDates = Array.from({ length: endDate }, (_, i) => i + 1);
-  const dates = [...emptyDates, ...filledDates];
-  const [clickFlag, setClickFlag] = useState(false);
   const handleDateClick = (e: React.MouseEvent<Element, MouseEvent>): void => {
     const targetDate = Number(e.currentTarget.textContent);
     if (!checkIn.month) {
@@ -61,6 +34,7 @@ const MonthlyCalendar: React.FunctionComponent<MonthProps> = ({
     }
     console.log(year, month);
   };
+
   const calendarTemplate = dates.map((date, idx) => (
     <div
       className={checkValidDate(date)}
